+++ test/_env/srv/handlers/main.js
@@ -177,8 +177,8 @@ module.exports = (srv) => {
     error.details = [
       {
         code: "ERR02-transition",
-        target: `Items(ID=2b750773-bb1b-4565-8a33-79c99440e4e8,IsActiveEntity=false)/description`,
-        ["@Common.additionalTargets"]: [`Items(ID=2b750773-bb1b-4565-8a33-79c99440e4e8,IsActiveEntity=false)/name`],
+        target: "Items(ID=2b750773-bb1b-4565-8a33-79c99440e4e8,IsActiveEntity=false)/description",
+        ["@Common.additionalTargets"]: ["Items(ID=2b750773-bb1b-4565-8a33-79c99440e4e8,IsActiveEntity=false)/name"],
         message: "Error details",
         "@Common.numericSeverity": 4,
         "@Core.ContentID": "1",
@@ -199,8 +199,8 @@ module.exports = (srv) => {
 
     const info2 = new Error("This is another warning");
     info2.code = "WARN02";
-    info2.target = `Items(ID=2b750773-bb1b-4565-8a33-79c99440e4e8,IsActiveEntity=false)/description`;
-    info2["@Common.additionalTargets"] = [`Items(ID=2b750773-bb1b-4565-8a33-79c99440e4e8,IsActiveEntity=false)/name`];
+    info2.target = "Items(ID=2b750773-bb1b-4565-8a33-79c99440e4e8,IsActiveEntity=false)/description";
+    info2["@Common.additionalTargets"] = ["Items(ID=2b750773-bb1b-4565-8a33-79c99440e4e8,IsActiveEntity=false)/name"];
     info2.message = "Another Warning occurred";
     info2.numericSeverity = 3;
     info2["@Core.ContentID"] = "2";
@@ -305,6 +305,40 @@ module.exports = (srv) => {
     ];
   });
 
+  srv.on("unboundErrorAction", async (req) => {
+    const error = new Error("An error occurred");
+    error.code = "ERR01";
+    switch (req.data.text) {
+      case "default":
+      default:
+        error.target = "name";
+        break;
+      case "unchecked":
+        error.target = "_xXx123";
+        break;
+      case "relative":
+        error.target = "Header(ID=1b750773-bb1b-4565-8a33-79c99440e4e8,IsActiveEntity=false)/name";
+        break;
+      case "guid":
+        error.target = "Header(ID=guid'1b750773-bb1b-4565-8a33-79c99440e4e8',IsActiveEntity=false)/name";
+        break;
+      case "absolute":
+        error.target = "/Header(ID=1b750773-bb1b-4565-8a33-79c99440e4e8,IsActiveEntity=false)/name";
+        break;
+      case "transient":
+        error.target = "/#TRANSIENT#/Header(ID=1b750773-bb1b-4565-8a33-79c99440e4e8,IsActiveEntity=false)/name";
+        break;
+      case "invalid":
+        error.target = "Header2(ID=1b750773-bb1b-4565-8a33-79c99440e4e8,IsActiveEntity=false)/name2";
+        break;
+    }
+    error.message = "An error occurred";
+    error.severity = 4;
+    error["@Common.numericSeverity"] = 4;
+    error["@Core.ContentID"] = "1";
+    req.error(error);
+  });
+
   /* Unbound Function */
 
   srv.on("unboundFunction", async (req) => {
@@ -393,9 +427,9 @@ module.exports = (srv) => {
   srv.on("unboundErrorFunction", async (req) => {
     const error = new Error("An error occurred");
     error.code = "ERR01";
-    error.target = `Header(ID=1b750773-bb1b-4565-8a33-79c99440e4e8,IsActiveEntity=false)/name`;
+    error.target = "Header(ID=1b750773-bb1b-4565-8a33-79c99440e4e8,IsActiveEntity=false)/name";
     error["@Common.additionalTargets"] = [
-      `Header(ID=1b750773-bb1b-4565-8a33-79c99440e4e8,IsActiveEntity=false)/description`,
+      "Header(ID=1b750773-bb1b-4565-8a33-79c99440e4e8,IsActiveEntity=false)/description",
     ];
     error.message = "An error occurred";
     error.severity = 4;
@@ -404,9 +438,10 @@ module.exports = (srv) => {
     error.details = [
       {
         code: "ERR02-transition",
-        target: `Header(ID=1b750773-bb1b-4565-8a33-79c99440e4e8,IsActiveEntity=false)/Items(ID=2b750773-bb1b-4565-8a33-79c99440e4e8,IsActiveEntity=false)/description`,
+        target:
+          "Header(ID=1b750773-bb1b-4565-8a33-79c99440e4e8,IsActiveEntity=false)/Items(ID=2b750773-bb1b-4565-8a33-79c99440e4e8,IsActiveEntity=false)/description",
         ["@Common.additionalTargets"]: [
-          `Header(ID=1b750773-bb1b-4565-8a33-79c99440e4e8,IsActiveEntity=false)/Items(ID=2b750773-bb1b-4565-8a33-79c99440e4e8,IsActiveEntity=false)/name`,
+          "Header(ID=1b750773-bb1b-4565-8a33-79c99440e4e8,IsActiveEntity=false)/Items(ID=2b750773-bb1b-4565-8a33-79c99440e4e8,IsActiveEntity=false)/name",
         ],
         message: "Error details",
         "@Common.numericSeverity": 4,
@@ -418,9 +453,9 @@ module.exports = (srv) => {
   srv.on("unboundWarningFunction", async (req) => {
     const info1 = new Error("This is a warning");
     info1.code = "WARN01";
-    info1.target = `Header(ID=1b750773-bb1b-4565-8a33-79c99440e4e8,IsActiveEntity=false)/name`;
+    info1.target = "Header(ID=1b750773-bb1b-4565-8a33-79c99440e4e8,IsActiveEntity=false)/name";
     info1["@Common.additionalTargets"] = [
-      `Header(ID=1b750773-bb1b-4565-8a33-79c99440e4e8,IsActiveEntity=false)/description`,
+      "Header(ID=1b750773-bb1b-4565-8a33-79c99440e4e8,IsActiveEntity=false)/description",
     ];
     info1.message = "An Warning occurred";
     info1.numericSeverity = 3;
@@ -429,9 +464,10 @@ module.exports = (srv) => {
 
     const info2 = new Error("This is another warning");
     info2.code = "WARN02";
-    info2.target = `Header(ID=1b750773-bb1b-4565-8a33-79c99440e4e8,IsActiveEntity=false)/Items(ID=2b750773-bb1b-4565-8a33-79c99440e4e8,IsActiveEntity=false)/description`;
+    info2.target =
+      "Header(ID=1b750773-bb1b-4565-8a33-79c99440e4e8,IsActiveEntity=false)/Items(ID=2b750773-bb1b-4565-8a33-79c99440e4e8,IsActiveEntity=false)/description";
     info2["@Common.additionalTargets"] = [
-      `Header(ID=1b750773-bb1b-4565-8a33-79c99440e4e8,IsActiveEntity=false)/Items(ID=2b750773-bb1b-4565-8a33-79c99440e4e8,IsActiveEntity=false)/name`,
+      "Header(ID=1b750773-bb1b-4565-8a33-79c99440e4e8,IsActiveEntity=false)/Items(ID=2b750773-bb1b-4565-8a33-79c99440e4e8,IsActiveEntity=false)/name",
     ];
     info2.message = "Another Warning occurred";
     info2.numericSeverity = 3;